+++ design-system/textarea/examples/constellation/text-area-default.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import TextArea from '../../src';
+
+export default () => (
+  <div css={{ '> *': { margin: '24px 0' } }}>
+    <TextArea
+      resize="auto"
+      maxHeight="20vh"
+      name="area"
+      defaultValue="Add a message here"
+    />
+  </div>
+);